Make EdgeBag implement java.util.Map.
EdgeBag is a map-like structure that has been custom built for the use-case
I was working on when I built the initial suffix tree, which was aiming at
indexing ASCII text on limited-memory machines.

This implementation would not work for non-ASCII character and is less
general than using a Character map.
Having EdgeBag extend Map will allow to refactor Node and
GeneralizedSuffixTree to use a user-selected Map implementation to store
the association between characters and edges.